requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.0.6
    # via kombu
asgiref==3.4.1
    # via django
backcall==0.2.0
    # via ipython
billiard==3.6.4.0
    # via celery
brotlipy==0.7.0
    # via -r requirements.in
celery[redis]==5.1.2
    # via
    #   -r requirements.in
    #   celery-redbeat
celery-redbeat==2.0.0
    # via -r requirements.in
certifi==2021.5.30
    # via sentry-sdk
cffi==1.14.6
    # via brotlipy
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
decorator==5.1.0
    # via ipython
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20200713
    # via django-import-export
dj-database-url==0.5.0
    # via -r requirements.in
django==3.2.7
    # via
    #   -r requirements.in
    #   django-import-export
    #   django-js-reverse
    #   django-log-request-id
    #   django-model-utils
    #   djangorestframework
django-debreach==2.0.1
    # via -r requirements.in
django-import-export==2.5.0
    # via -r requirements.in
django-js-reverse==0.9.1
    # via -r requirements.in
django-log-request-id==1.6.0
    # via -r requirements.in
django-model-utils==4.1.1
    # via -r requirements.in
django-webpack-loader==1.3.0
    # via -r requirements.in
djangorestframework==3.12.4
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
gunicorn==20.1.0
    # via -r requirements.in
ipython==7.27.0
    # via -r requirements.in
jedi==0.18.0
    # via ipython
kombu==5.1.0
    # via celery
markuppy==1.14
    # via tablib
matplotlib-inline==0.1.3
    # via ipython
odfpy==1.4.1
    # via tablib
openpyxl==3.0.8
    # via tablib
parso==0.8.2
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.20
    # via
    #   click-repl
    #   ipython
psutil==5.8.0
    # via -r requirements.in
psycopg2==2.9.1
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pycparser==2.20
    # via cffi
pygments==2.10.0
    # via ipython
python-dateutil==2.8.2
    # via celery-redbeat
python-decouple==3.4
    # via -r requirements.in
pytz==2021.1
    # via
    #   celery
    #   django
pyyaml==5.4.1
    # via tablib
redis==3.5.3
    # via
    #   celery
    #   celery-redbeat
sentry-sdk==1.3.1
    # via -r requirements.in
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
sqlparse==0.4.2
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.0.0
    # via django-import-export
tenacity==8.0.1
    # via celery-redbeat
traitlets==5.1.0
    # via
    #   ipython
    #   matplotlib-inline
urllib3==1.26.6
    # via sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
whitenoise==5.3.0
    # via -r requirements.in
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
